Refactor: Use schema provider registry from core
